use the same timestamp type for input and output

+++ src/eyefi.h
@@ -10,7 +10,7 @@ struct rns__StartSessionResponse {
     std::string credential;
     std::string snonce;
     int transfermode;
-    unsigned int transfermodetimestamp;
+    long transfermodetimestamp;
     bool upsyncallowed;
 };
